Move OrderMetadataService
