+++ api/src/main/java/org/openmrs/module/emrapi/encounter/OrderMetadataService.java
@@ -11,7 +11,7 @@
  *
  * Copyright (C) OpenMRS, LLC.  All Rights Reserved.
  */
-package org.openmrs.module.emrapi.encounter.service;
+package org.openmrs.module.emrapi.encounter;
 
 import org.openmrs.Concept;
 import org.openmrs.OrderFrequency;
