+++ api/src/test/java/org/openmrs/module/emrapi/encounter/mapper/OpenMRSDrugOrderMapper1_12Test.java
@@ -34,7 +34,7 @@ import org.openmrs.api.ConceptService;
 import org.openmrs.api.OrderService;
 import org.openmrs.module.emrapi.encounter.builder.DrugOrderBuilder;
 import org.openmrs.module.emrapi.encounter.domain.EncounterTransaction;
-import org.openmrs.module.emrapi.encounter.service.OrderMetadataService;
+import org.openmrs.module.emrapi.encounter.OrderMetadataService;
 
 import java.text.ParseException;
 import java.util.Date;
